Enable -Wconversion

+++ zz.h
@@ -37,7 +37,7 @@ typedef struct {
 } zz_t;
 
 zz_err zz_init(zz_t *u);
-zz_err zz_resize(size_t size, zz_t *u);
+zz_err zz_resize(int64_t size, zz_t *u);
 void zz_clear(zz_t *u);
 
 typedef enum {
